📝 Moving the tests to sample app
