+++ android/build.gradle
@@ -36,6 +36,4 @@ android {
 dependencies {
     implementation 'com.instabug.library:instabug:8.2.2'
     testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:1.10.19'
-    implementation 'org.mockito:mockito-core:1.10.19'
 }
