+++ example/android/app/build.gradle
@@ -58,4 +58,6 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    testImplementation 'org.mockito:mockito-core:1.10.19'
+    implementation 'org.mockito:mockito-core:1.10.19'
 }
