+++ example/android/app/src/test/java/com/instabug/instabugflutterexample/InstabugFlutterPluginTest.java
@@ -1,9 +1,6 @@
-package com.instabug.instabugflutter;
+package com.instabug.instabugflutterexample;
 
 import org.junit.Test;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 
 
 public class InstabugFlutterPluginTest {
